Demo fixes

+++ domainmodels/src/main/java/com/example/domainmodels/ServerStatusDTO.kt
@@ -7,9 +7,9 @@ data class ServerStatusDTO(
     @Json(name="ok")
     val ok: Boolean,
     @Json(name="uptime")
-    val uptime: Int,
+    val uptime: Float,
 ) {
     companion object {
-        val EMPTY = ServerStatusDTO(ok = false, uptime = 0)
+        val EMPTY = ServerStatusDTO(ok = false, uptime = 0f)
     }
 }
